+++ packages/runtime/src/runtime.ts
@@ -20,7 +20,6 @@ export type Options = {
 
   timeout?: number; // this is timeout used per job, not per workflow
   strict?: boolean; // Be strict about handling of state returned from jobs
-  deleteConfiguration?: boolean;
 
   // Treat state as immutable (likely to break in legacy jobs)
   immutableState?: boolean;
@@ -36,6 +35,9 @@ export type Options = {
 
   // inject globals into the environment
   globals?: any;
+
+  // all listed props will be removed from the state object at the end of a job
+  statePropsToRemove?: string[];
 };
 
 const defaultState = { data: {}, configuration: {} };
@@ -56,8 +58,8 @@ const run = (
   if (!opts.hasOwnProperty('strict')) {
     opts.strict = true;
   }
-  if (!opts.hasOwnProperty('deleteConfiguration')) {
-    opts.deleteConfiguration = true;
+  if (!opts.hasOwnProperty('statePropsToRemove')) {
+    opts.statePropsToRemove = ['configuration'];
   }
 
   // TODO the plan doesn't have an id, should it be given one?