+++ packages/runtime/test/runtime.test.ts
@@ -117,7 +117,7 @@ test('resolve a credential', async (t) => {
 
   const options = {
     strict: false,
-    deleteConfiguration: false,
+    statePropsToRemove: [],
     callbacks: {
       resolveCredential: async () => ({ password: 'password1' }),
     },