+++ integration-tests/worker/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @openfn/integration-tests-worker
 
+## 1.0.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @openfn/ws-worker@0.5.0
+
+## 1.0.28
+
+### Patch Changes
+
+- Updated dependencies [5c45e1e]
+  - @openfn/engine-multi@0.2.6
+  - @openfn/lightning-mock@1.1.8
+  - @openfn/ws-worker@0.4.1
+
+## 1.0.27
+
+### Patch Changes
+
+- Updated dependencies [f228fd5]
+  - @openfn/ws-worker@0.4.0
+  - @openfn/engine-multi@0.2.5
+  - @openfn/lightning-mock@1.1.7
+
+## 1.0.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @openfn/ws-worker@0.3.2
+
 ## 1.0.25
 
 ### Patch Changes